+++ src/components/form/spot.form.tsx
@@ -1,15 +1,15 @@
 "use client";
 
 /* LIBRAIRIES */
-import React from "react";
+import React, { useEffect } from "react";
 import { useForm, FormProvider, useFieldArray } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
 import * as yup from "yup";
-import { toast } from "sonner";
 import { Spin } from "antd";
 
 /* ACTIONS */
-import { CREATE_SPOT } from "@/libs/actions";
+import { CREATE_SPOT, UPDATE_SPOT } from "@/libs/actions";
+import { spotSchema } from "@/libs/yup";
 
 /* STORES */
 import { useSpots, useActivities } from "@/context/store";
@@ -22,29 +22,6 @@ import { Input, CheckboxInput, Textarea } from "@/components/Inputs";
 import Modal from "@/components/Modal";
 import ToasterAction from "@/components/ToasterAction";
 
-
-/* Validation schema */
-const baseSchema = yup.object().shape({
-  name: yup.string().required("Le champ name est requis"),
-  description: yup.string(),
-  gpsCoordinates: yup.string().required("Le champ gpsCoordinates est requis"),
-  photo: yup.string(),
-  half_day: yup.boolean(),
-  full_day: yup.boolean(),
-  max_OfPeople: yup
-    .number()
-    .positive()
-    .integer()
-    .required("Le champ max_OfPeople est requis"),
-  min_OfPeople: yup
-    .number()
-    .positive()
-    .integer()
-    .required("Le champ min_OfPeople est requis"),
-  meetingPoint: yup.string(),
-  estimatedDuration: yup.string(),
-});
-
 const practicedActivitiesSchema = yup.array().of(
   yup.object().shape({
     activityId: yup.string().required("Le champ activityId est requis"),
@@ -53,7 +30,7 @@ const practicedActivitiesSchema = yup.array().of(
 );
 
 const createDynamicSchema = (fields: Array<any>) => {
-  return baseSchema.concat(
+  return spotSchema.concat(
     yup.object().shape({
       practicedActivities: practicedActivitiesSchema.min(
         1,
@@ -80,17 +57,24 @@ export type TSpotForm = {
   estimatedDuration: string | undefined;
 };
 
-export default function CreateSpotForm({
+export  function SpotForm({
+  spotData,
   isOpen,
   onClose,
 }: {
+  spotData?: ISpot;
   isOpen: boolean;
   onClose: () => void;
 }) {
+  const isUpdate = !!spotData;
+
   const updateSpot = useSpots((state) => state.updateSpots);
   const activities = useActivities((state) => state.Activities);
   const methods = useForm<TSpotForm>({
     resolver: yupResolver(createDynamicSchema([])),
+    defaultValues: {
+      ...spotData,
+    },
   });
 
   const { fields, append, remove } = useFieldArray({
@@ -99,15 +83,17 @@ export default function CreateSpotForm({
   });
 
   const onSubmit = async (data: TSpotForm) => {
-    const result = await CREATE_SPOT(data as ISpot);
+    const result = isUpdate
+      ? await UPDATE_SPOT(spotData!._id, data as ISpot)
+      : await CREATE_SPOT(data as ISpot);
+
     if (result.success) {
       if (result.data) {
         updateSpot(result.data);
       }
-      reset();
-      onClose();
-    } 
-    ToasterAction({result , defaultMessage: "Lieu créé avec succès"})
+      handleClose();
+    }
+    ToasterAction({ result, defaultMessage: isUpdate ? "Lieu modifié avec succès" : "Lieu créé avec succès" });
   };
 
   const {
@@ -116,75 +102,68 @@ export default function CreateSpotForm({
     formState: { isSubmitting, errors },
   } = methods;
 
+  useEffect(() => {
+    if (spotData) {
+      reset({
+        ...spotData,
+      });
+    }
+  }, [spotData, reset]);
+
+  const handleClose = () => {
+    reset();
+    onClose();
+  };
+
   return (
-    <Modal isOpen={isOpen} onClose={onClose}>
+    <Modal isOpen={isOpen} onClose={handleClose}>
       <FormProvider {...methods}>
-        <form
-          onSubmit={handleSubmit(onSubmit)}
-          className="flex flex-col justify-center items-center gap-6"
-        >
+        <form onSubmit={handleSubmit(onSubmit)} className="flex flex-col justify-center items-center gap-6">
           {/* HEADER */}
           <div className="flex flex-col justify-center items-center gap-1">
-            <h2 className="text-2xl font-bold">Créer un lieu</h2>
+            <h2 className="text-2xl font-bold">{isUpdate ? "Modifier le lieu" : "Créer un lieu"}</h2>
           </div>
 
           {/* FORM */}
           <div className="flex flex-col items-center gap-2 border-2 rounded-md border-sky-500 w-full p-2">
-            <p className="text-sky-500 text-xl font-bold">
-              Informations du lieu
-            </p>
+            <p className="text-sky-500 text-xl font-bold">Informations du lieu</p>
 
             <Input name="name" type="text" label="Nom" />
-            <Textarea
-              name="description"
-              label="Description"
-              placeholder="Description"
-            />
+            <Textarea name="description" label="Description" placeholder="Description" />
             <div className="flex flex-col md:flex-row gap-4">
               <Input
                 name="gpsCoordinates"
                 type="text"
                 label="Coordonnées GPS"
-                className=""
+                
                 placeholder="Exemple: 48.8584, 2.2945"
               />
               <Input
                 name="meetingPoint"
                 type="text"
                 label="Point de rencontre"
-                className=""
+                
                 placeholder="Exemple: 48.8584, 2.2945"
               />
             </div>
 
-            <Input
-              name="estimatedDuration"
-              type="text"
-              label="Durée estimée"
-              className=""
-            />
+            <Input name="estimatedDuration" type="text" label="Durée estimée"  />
           </div>
 
           <div className="flex flex-col items-center gap-1 border-2 rounded-md border-sky-500 w-full p-2">
-            <p className="text-sky-500 text-xl font-bold">
-              Activités pratiquées
-            </p>
+            <p className="text-sky-500 text-xl font-bold">Activités pratiquées</p>
 
             {/* message d'erreur si aucune activité n'est sélectionnée */}
             {errors.practicedActivities && (
-              <p className="text-red-500">
-                Au moins une activité doit être sélectionnée
-              </p>
+              <p className="text-red-500">Au moins une activité doit être sélectionnée</p>
             )}
-            <div className="grid grid-cols-2 gap-4 justify-around  ">
+            <div className="grid grid-cols-2 gap-4 justify-around">
               {activities.map((activity) => (
                 <CheckboxInput
                   key={activity._id}
                   name={`practicedActivities.${activity._id}`}
                   label={activity.name}
-                  checked={fields.some(
-                    (field) => field.activityId === activity._id
-                  )}
+                  checked={fields.some((field) => field.activityId === activity._id)}
                   onChange={(e) => {
                     if (e.target.checked) {
                       append({
@@ -192,9 +171,7 @@ export default function CreateSpotForm({
                         activityName: activity.name,
                       });
                     } else {
-                      const index = fields.findIndex(
-                        (field) => field.activityId === activity._id
-                      );
+                      const index = fields.findIndex((field) => field.activityId === activity._id);
                       if (index !== -1) remove(index);
                     }
                   }}
@@ -206,34 +183,16 @@ export default function CreateSpotForm({
           <div className="flex flex-col items-center gap-1 border-2 rounded-md border-sky-500 w-full p-2">
             <p className="text-sky-500 text-xl font-bold">Disponibilité</p>
             <div className="flex flex-col md:flex-row gap-4 w-full justify-center">
-              <CheckboxInput
-                name="half_day"
-                label="Demi-journée"
-                className=""
-              />
-              <CheckboxInput
-                name="full_day"
-                label="Journée complète"
-                className=""
-              />
+              <CheckboxInput name="half_day" label="Demi-journée"  />
+              <CheckboxInput name="full_day" label="Journée complète"  />
             </div>
           </div>
 
           <div className="flex flex-col items-center gap-1 border-2 rounded-md border-sky-500 w-full p-2">
             <p className="text-sky-500 text-xl font-bold">Capacité</p>
             <div className="flex flex-col md:flex-row gap-4 w-full justify-center">
-              <Input
-                name="max_OfPeople"
-                type="number"
-                label="Nombre maximum de personnes"
-                className=""
-              />
-              <Input
-                name="min_OfPeople"
-                type="number"
-                label="Nombre minimum de personnes"
-                className=""
-              />
+              <Input name="max_OfPeople" type="number" label="Nombre maximum de personnes"  />
+              <Input name="min_OfPeople" type="number" label="Nombre minimum de personnes"  />
             </div>
           </div>
 
@@ -244,7 +203,7 @@ export default function CreateSpotForm({
               className="bg-orange-500 hover:bg-orange-600 transition-all duration-300 text-white w-fit mx-auto p-3 rounded-md flex items-center justify-center min-w-[70px] min-h-[40px] disabled:opacity-80 disabled:cursor-not-allowed"
               disabled={isSubmitting}
             >
-              {isSubmitting ? <Spin size="default" /> : "Créer"}
+              {isSubmitting ? <Spin size="default" /> : isUpdate ? "Modifier" : "Créer"}
             </button>
           </div>
         </form>
