+++ src/libs/actions/customerSession.actions.ts
@@ -8,6 +8,7 @@ import { connectDB, disconnectDB } from "@/libs/database/mongodb";
 /* Models */
 import CustomerSession from "@/libs/database/models/CustomerSession";
 import Session from "@/libs/database/models/Session";
+import { customerSessionSchema } from "@/libs/yup";
 
 /* Types */
 import {
@@ -23,36 +24,7 @@ import {
 import { GET_SERVER_SESSION_WITH_DETAILS } from "@/libs/actions/sessionWithDetail.actions";
 
 
-/*
- * CustomerSession schema
- */
-const customerSessionSchema = yup.object().shape({
-  sessionId: yup.string().required("Le champ sessionId est requis"),
-  date: yup.date().required("Le champ date est requis"),
-  status: yup
-    .string()
-    .required("Le champ status est requis")
-    .oneOf(["Validated", "Canceled", "Waiting"]),
-  typeOfReservation: yup
-    .string()
-    .required("Le champ typeOfReservation est requis"),
-  number_of_people: yup
-    .number()
-    .required("Le champ number_of_people est requis"),
-  last_name: yup.string().required("Le champ last_name est requis"),
-  first_names: yup.string().required("Le champ first_names est requis"),
-  email: yup
-    .string()
-    .email("L'email est invalide")
-    .required("Le champ email est requis"),
-  phone: yup.string().required("Le champ phone est requis"),
-  people_list: yup.array().of(
-    yup.object().shape({
-      size: yup.string().required("Le champ size est requis"),
-      weight: yup.string().required("Le champ weight est requis"),
-    })
-  ),
-});
+
 
 /*
  * Validate the customerSession
